ECA-34 refactor: rename class Product and add field in IProduct
