+++ src/components/catalog/index.ts
@@ -1,5 +1,5 @@
 import createElement from '../../dom-helper/create-element';
-import Product from '../product';
+import ProductCard from '../product-card';
 import './catalog.scss';
 
 import getAllProducts from '../../controller/get-all-products';
@@ -7,12 +7,12 @@ import getAllProducts from '../../controller/get-all-products';
 export default class Catalog {
   private container = createElement('section', { class: 'catalog' });
 
-  private products: Product[] | null = null;
+  private products: ProductCard[] | null = null;
 
   private init(): void {
     getAllProducts()
       .then((productsResponse) => {
-        this.products = productsResponse.map((product) => new Product(product));
+        this.products = productsResponse.map((product) => new ProductCard(product));
         this.products?.forEach((product) => this.container.append(product.render()));
       })
       .catch((err) => {
