+++ src/components/product-card/product-card.scss
