+++ src/types/index.ts
@@ -1,3 +1,3 @@
-import IProduct from '../components/product';
+import IProduct from '../components/product-card';
 
 export { IProduct };
