+++ src/types/product.ts
@@ -1,5 +1,6 @@
 export default interface IProduct {
   id: string;
+  variantId?: number;
   name: string;
   description: string;
   price: string;
